finish staff controller

+++ Modules/Staff/Staff.Controller.js
@@ -2,6 +2,7 @@ import asyncHandler from "express-async-handler";
 import staffModel from "../../Database/Models/Staff.Model.js";
 import ApiError from "../../utils/apiError.js";
 import adminModel from "../../Database/Models/Admin.Model.js";
+import jwt from "jsonwebtoken";
 
 const getAllStaff = asyncHandler(async (req, res) => {
   // to add pagination
@@ -17,42 +18,48 @@ const getAllStaff = asyncHandler(async (req, res) => {
 const getOneStaff = asyncHandler(async (req, res, next) => {
   const { id } = req.params;
   const staff = await staffModel.findById(id);
-  if (!staff) {
-    // res.status(404).json({ message: `no staff for this id ${id}` });
-    return next(new ApiError(`no staff for this id ${id}`, 404));
-  } else {
+  if (staff) {
     res.status(200).json({ status: "success", data: staff });
+  } else {
+    return next(new ApiError(`no staff for this id ${id}`, 404));
   }
 });
 
 const addStaff = asyncHandler(async (req, res, next) => {
-  const { ssn, name, jobTitle, age, salary, phone, email, gender } = req.body;
-  const creatorId = req.decodedToken.id;
+  const decoded = jwt.verify(req.headers.token, "bl7 5ales");
+  const { id: creatorId } = decoded;
   const admin = await adminModel.findById(creatorId);
   if (admin) {
-    const staff = await staffModel.create({
-      createdBy: creatorId,
-      ...req.body,
-    });
-    res.status(201).json({ status: "success", data: staff });
+    const newuser = await staffModel.insertMany([
+      { ...req.body, createdBy: creatorId },
+    ]);
+    res
+      .status(200)
+      .json({ status: "Success", message: "User added", data: newuser });
   } else {
-    return next(new ApiError("invaild token", 401));
+    return next(new ApiError("Unauthorized", 401));
   }
 });
 
 const updataStaff = async (req, res, next) => {
   try {
-    const creatorId = req.decodedToken.id;
+    const decoded = jwt.verify(req.headers.token, "bl7 5ales");
+    const { id: creatorId } = decoded;
     const admin = await adminModel.findById(creatorId);
     if (admin) {
       const { id } = req.params;
-      const { name, salary, jobTitle } = req.body;
-      const staff = await staffModel.findOneAndUpdate(
-        { _id: id },
-        { name, salary, jobTitle },
-        { new: true }
-      );
-      res.status(200).json({ status: "success", data: staff });
+      const isFound = await staffModel.findById(id);
+      if (isFound) {
+        const { name, salary, jobTitle } = req.body;
+        const staff = await staffModel.findOneAndUpdate(
+          { _id: id },
+          { name, salary, jobTitle },
+          { new: true }
+        );
+        res.status(200).json({ status: "success", data: staff });
+      } else {
+        res.status(404).json({ status: "Fail", message: "Staff not found" });
+      }
     } else {
       return next(new ApiError("no staff for this id ${id}", 404));
     }
@@ -63,12 +70,12 @@ const updataStaff = async (req, res, next) => {
 
 const deleteStaff = asyncHandler(async (req, res, next) => {
   const { id } = req.params;
-  const staff = await staffModel.findByIdAndDelete(id);
-  if (!staff) {
-    return next(new ApiError(`no staff for this id ${id}`, 404));
-  } else {
+  const isFound = await staffModel.findById(id);
+  if (isFound) {
+    const staff = await staffModel.findByIdAndDelete(id);
     res.status(204).json({ status: "success" });
   }
+  return next(new ApiError(`no staff for this id ${id}`, 404));
 });
 
 export { getAllStaff, getOneStaff, addStaff, updataStaff, deleteStaff };
